feat: Add server address list to RPC app
This commit adds the server address list to the RPC app in the client.py file. The server address list is populated by making a request to the server and retrieving a list of addresses. This change improves the functionality of the RPC app by allowing it to connect to multiple servers.